adding tests for remove book, add book and adjusting remove book method

+++ Business logic/Services/BookServices.cs
@@ -67,6 +67,7 @@ namespace Bookly.Business_logic.Services
 
         public bool RemoveBook(int id)
         {
+            if (GetBookById(id) == null) return false;
             return _bookRepo.RemoveBook(id);
         }
